+++ src/blogs.tsx
@@ -2,6 +2,8 @@ import React, { useState } from "react";
 import "./blogs.css";
 import { Button, TextField } from "@mui/material";
 import { PostfixUnaryExpression } from "typescript";
+import AddCommentComp from "./comment";
+import { usePost } from "./hooks";
 // import ReactDOM  from "react-dom";
 // import Button from '@mui/material/Button'
 
@@ -18,122 +20,10 @@ type NewPost = {
   content: string;
 };
 
-type Comment = {
-  id: number;
-  author: string;
-  content: string;
-  createdAt: string;
-};
-
-type NewCommetn = {
-  author: string;
-  content: string;
-};
-
-// コメントコンポーネント
-const AddCommentComp = () => {
-  const [comments, setComments] = useState<Comment[]>([]);
-  const [newComment, setNewComment] = useState<NewCommetn>({
-    author: "",
-    content: "",
-  });
-  const AddComment = () => {
-    setComments([
-      ...comments,
-      {
-        id: comments.length + 1,
-        author: newComment.author,
-        content: newComment.content,
-        createdAt: new Date().toLocaleString(),
-      },
-    ]);
-    setNewComment({ author: "", content: "" });
-  };
-  return (
-    <div>
-      <TextField
-        id="postAuthor"
-        label="コメント投稿者名"
-        variant="outlined"
-        value={newComment.author}
-        onChange={(e) =>
-          setNewComment({
-            author: e.target.value,
-            content: newComment.content,
-          })
-        }
-      />
-      <br></br>
-      <TextField
-        id="postContent"
-        label="コメント内容"
-        variant="outlined"
-        value={newComment.content}
-        onChange={(e) =>
-          setNewComment({
-            author: newComment.author,
-            content: e.target.value,
-          })
-        }
-      />
-      <div className="addComment">
-        <Button
-          variant="contained"
-          onClick={AddComment}
-          disabled={newComment.author === "" || newComment.content === ""}
-        >
-          コメントする
-        </Button>
-      </div>
-      <div className="commnetList">
-        <ul>
-          {comments.map((comment, id) => (
-            <li key={id}>
-              {comment.author}:{comment.content}
-            </li>
-          ))}
-        </ul>
-      </div>
-    </div>
-  );
-};
 
 // ブログ全体のコンポーネント
 const Blogs = () => {
-  const [posts, setPosts] = useState<Post[]>([]);
-  const [newPost, setNewPost] = useState<NewPost>({ author: "", content: "" });
-
-  const PostSituation = newPost.author !== "" && newPost.content !== "";
-
-  const AddPost = () => {
-    if (!PostSituation) {
-      return;
-    }
-    setPosts([
-      ...posts,
-      {
-        id: posts.length + 1,
-        author: newPost.author,
-        content: newPost.content,
-        createdAt: new Date().toLocaleString(),
-        comments: [],
-      },
-    ]);
-    setNewPost({ author: "", content: "" });
-  };
-
-  //   const AddComment = () => {
-  //     setComments([
-  //       ...comments,
-  //       {
-  //         id: comments.length + 1,
-  //         author: newComment.author,
-  //         content: newComment.content,
-  //         createdAt: new Date().toLocaleString(),
-  //       },
-  //     ]);
-  //     setNewComment({author:"", content:""})
-  //   };
+  const {newPost, setNewPost,posts, setPosts, addPost} = usePost()
 
   return (
     <div>
@@ -163,7 +53,7 @@ const Blogs = () => {
       <div>
         <Button
           variant="contained"
-          onClick={AddPost}
+          onClick={addPost}
           disabled={newPost.author === "" || newPost.content === ""}
         >
           投稿
